+++ src/main/java/io/reactivex/rxjava3/internal/subscriptions/DeferredScalarSubscription.java
@@ -17,7 +17,6 @@ import org.reactivestreams.Subscriber;
 
 import io.reactivex.rxjava3.annotations.Nullable;
 
-
 /**
  * A subscription that signals a single value eventually.
  * <p>
@@ -75,7 +74,7 @@ public class DeferredScalarSubscription<T> extends BasicIntQueueSubscription<T> 
     public final void request(long n) {
         if (SubscriptionHelper.validate(n)) {
             for (;;) {
-                int state = ((Integer)(get())).intValue();
+                int state = get();
                 // if the any bits 1-31 are set, we are either in fusion mode (FUSED_*)
                 // or request has been called (HAS_REQUEST_*)
                 if ((state & ~NO_REQUEST_HAS_VALUE) != 0) {
